Added support for reflecting on associations for class determination in collection_field.
This allows us to use association reflection when there is no item in the list
of already-selected items to try and extract the class name from. That means we
can now correctly look up the full list for the collection field even when there
is nothing pre-selected -- useful for new fields in general.